Move the rest of the UI classes
