+++ Sidekick.sln
@@ -19,11 +19,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sidekick.Core", "src\Sideki
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sidekick", "src\Sidekick\Sidekick.csproj", "{DBDA2C98-1E9D-4E5B-866B-297FF28B824E}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sidekick.UI", "src\Sidekick.UI\Sidekick.UI.csproj", "{F8DCDD39-7D3D-4610-A677-062410F89779}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sidekick.Localization", "src\Sidekick.Localization\Sidekick.Localization.csproj", "{C7A61A43-9474-415E-84AD-381E45003C6F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sidekick.Business.Tests", "Sidekick.Business.Tests\Sidekick.Business.Tests.csproj", "{B7235282-296F-473E-80BC-47D620D80CA0}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sidekick.Business.Tests", "Sidekick.Business.Tests\Sidekick.Business.Tests.csproj", "{B7235282-296F-473E-80BC-47D620D80CA0}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{B4E782B6-CF03-4448-81D1-3E308D20BF2B}"
 EndProject
@@ -45,10 +43,6 @@ Global
 		{DBDA2C98-1E9D-4E5B-866B-297FF28B824E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DBDA2C98-1E9D-4E5B-866B-297FF28B824E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DBDA2C98-1E9D-4E5B-866B-297FF28B824E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F8DCDD39-7D3D-4610-A677-062410F89779}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F8DCDD39-7D3D-4610-A677-062410F89779}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F8DCDD39-7D3D-4610-A677-062410F89779}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F8DCDD39-7D3D-4610-A677-062410F89779}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C7A61A43-9474-415E-84AD-381E45003C6F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C7A61A43-9474-415E-84AD-381E45003C6F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C7A61A43-9474-415E-84AD-381E45003C6F}.Release|Any CPU.ActiveCfg = Release|Any CPU
