+++ src/Sidekick.UI/Sidekick.UI.csproj
@@ -18,4 +18,8 @@
     <ProjectReference Include="..\Sidekick.Localization\Sidekick.Localization.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Views\" />
+  </ItemGroup>
+
 </Project>
