+++ src/Sidekick/Sidekick.csproj
@@ -82,7 +82,6 @@
   <ItemGroup>
     <ProjectReference Include="..\Sidekick.Business\Sidekick.Business.csproj" />
     <ProjectReference Include="..\Sidekick.Localization\Sidekick.Localization.csproj" />
-    <ProjectReference Include="..\Sidekick.UI\Sidekick.UI.csproj" />
   </ItemGroup>
 
   <ItemGroup>
