+++ src/Sidekick/Views/SplashScreen/SplashScreenView.xaml
@@ -1,5 +1,5 @@
 <Window
-    x:Class="Sidekick.Views.SplashScreen"
+    x:Class="Sidekick.Views.SplashScreen.SplashScreenView"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:loc="clr-namespace:Sidekick.Views"
